Refactoring: sse_manager is separated
